Add all feature v1

+++ mcrs/lm_modules/llama.py
@@ -10,11 +10,16 @@ class LLAMA_MODEL:
         self.attn_implementation = attn_implementation
         self.lm, self.tokenizer = self._load_model()
         self.lm.eval()
-        self.lm.to(self.device).to(self.dtype)
+        self.lm.to(self.device)
 
     def _load_model(self):
         tokenizer = AutoTokenizer.from_pretrained(self.model_name, padding_side="left")
-        lm = AutoModelForCausalLM.from_pretrained(self.model_name, attn_implementation=self.attn_implementation, dtype=self.dtype)
+        # `torch_dtype` is the correct kwarg for from_pretrained (not `dtype`)
+        lm = AutoModelForCausalLM.from_pretrained(
+            self.model_name,
+            attn_implementation=self.attn_implementation,
+            torch_dtype=self.dtype,
+        )
         return lm, tokenizer
 
     def _format_chat_history(self, sys_prompt, chat_history: list, recommend_item: str):
